Rename image/interface files
